rename luckybet
